+++ .babelrc.js
@@ -1,13 +1,16 @@
 module.exports = {
-  plugins: ['@babel/plugin-proposal-optional-chaining'],
+  plugins: [
+    '@babel/proposal-nullish-coalescing-operator',
+    '@babel/proposal-optional-chaining',
+  ],
   presets: [
     [
-      '@babel/preset-env',
+      '@babel/env',
       {
         modules: process.env.NODE_ENV === 'test' && 'commonjs',
         targets: {node: '6'},
       },
     ],
-    '@babel/preset-flow',
+    '@babel/flow',
   ],
 };
