+++ .lintstagedrc.yml
@@ -1,3 +1,9 @@
 linters:
- "*.{json,md}": ["prettier --write", "git add"]
- "*.{js}": ["prettier --write", "eslint --fix", "git add", "jest --config ./scripts/jest/config.source.js --bail --findRelatedTests"]
+  '*.{json,md,yml}': ['prettier --write', 'git add']
+  '*.js':
+    [
+      'prettier --write',
+      'eslint --fix',
+      'git add',
+      'jest --config ./scripts/jest/config.source.js --bail --findRelatedTests',
+    ]
