+++ package.json
@@ -10,18 +10,19 @@
   },
   "description": "Load configuration from environmental variables and files.",
   "devDependencies": {
-    "@babel/core": "^7.0.0-beta.51",
-    "@babel/plugin-proposal-optional-chaining": "^7.0.0-beta.51",
-    "@babel/preset-env": "^7.0.0-beta.51",
-    "@babel/preset-flow": "^7.0.0-beta.51",
+    "@babel/core": "7.0.0-rc.1",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "7.0.0-rc.1",
+    "@babel/plugin-proposal-optional-chaining": "7.0.0-rc.1",
+    "@babel/preset-env": "7.0.0-rc.1",
+    "@babel/preset-flow": "7.0.0-rc.1",
     "@commitlint/cli": "^7.0.0",
     "@commitlint/config-conventional": "^7.0.0",
     "@wtg/scripts": "^1.1.0",
     "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "^8.2.4",
+    "babel-eslint": "9.0.0-beta.3",
     "babel-jest": "^23.0.0",
-    "eslint": "^5.0.0",
-    "eslint-config-airbnb-base": "^13.0.0",
+    "eslint": "^5.3.0",
+    "eslint-config-airbnb-base": "^13.1.0",
     "eslint-config-prettier": "^3.0.0",
     "eslint-plugin-flowtype": "^2.46.1",
     "eslint-plugin-import": "^2.9.0",
@@ -48,7 +49,7 @@
   "scripts": {
     "build": "wtg build",
     "clean": "rimraf coverage dist",
-    "lint": "flow && eslint . && prettier --l \"{__tests__,src}/**/*.js\"",
+    "lint": "flow && eslint . && prettier --l \"{__mocks__,__tests__,src}/**/*.js\"",
     "test": "yarn test:src && yarn build && yarn test:dist",
     "test:src": "jest --config scripts/jest/config.source.js",
     "test:dist": "jest --config scripts/jest/config.dist.js"
