+++ src/loadFileConfigSync.js
@@ -4,7 +4,7 @@ import type {FileConfig} from './types';
 
 export default function loadFileConfigSync(
   property: FileConfig,
-): {|error?: false | Error, value?: string|} {
+): {|error?: ?(false | Error), value?: ?string|} {
   const {encoding = 'utf8', filePath, required = false} = property;
   try {
     const value = readFileSync(filePath, encoding);
